[FIX] hr_appraisal: fix goal view
Before this commit, when a user doesn't have any hr right,
he can't see employee's picture.

With this commit, this issue is fixed and the user can access to the
public employee form when he clicks on this picture. And "mark as done"
option in three dot menu is invisible if the goal is already done.

task-4240631

closes odoo/enterprise#72497

X-original-commit: 723a40770170a55636265c17a16bf42b885636ae
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Mélanie Peyrat (mepe) <mepe@odoo.com>