[FIX] sale_stock_renting: check child locations for outgoing moves
**Current behavior:**
Using 2 step delivery and modifying the route so the pick
action's src location is a child of lot_stock will result in the
move not being found by the Procurement mechanism, which can
lead to inaccurate quantity being encoded on transfers created
from rental orders.

**Expected behavior:**
Transfer quantity should match sale order line quantity.

**Steps to reproduce:**
1. Enable 2-step delivery in the warehouse

2. In the 2-step delivery route, change the src location of the
pick action (pull rule) to be a sub-location of WH/STOCK

3. Enable rental transfers in settings

4. Create a new rental order, add a product, confirm the order

5. Add some quantity to the order line and save

6. Look at the generated out-going picking's moves -> the
demand is wrong

**Cause of the issue:**
The original move created when confirming the order, before
editing the order line quantity, is not found when the
additional move is created for the modified quantity.

Existing outgoing moves are expected to have a location_id
(source location) of WH/STOCK (warehouse.lot_stock_id), ours
will have a location_id of some sub-location.

**Fix:**
Check if the move's location_id is a child of lot_stock.

opw-4178446

closes odoo/enterprise#71821

X-original-commit: 88c4e6004674e68d06dd31d1e3c589d917cebb14
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>