[FIX] l10n_cl_edi{_boletas}: include parent when checking tax group in a branch
Steps to reproduce:
- Install Accounting and l10n_cl_edi_boletas
- Switch to a Chilean company (e.g. CL Company)
- Create a branch company for it
- Create a product with a tax from "ILA" group
- Switch to the branch company
- Create an invoice with the created product
- Confirm it

Issue:
A traceback is raised while trying to fetch some tax groups:
- self.env.ref(f'account.{cid}_tax_group_ila')
- self.env.ref(f'account.{cid}_tax_group_retenciones')
where {cid} is the id of the branch company.

Cause:
There is no tax group defined in the branch company.
The taxes and the tax groups are defined in the parent company.

Solution:
Do not raise an error if the XMLID cannot be found and also check
tax groups from parent companies as the taxes could come from them.
Use the "ref" method defined in "account.chart.template" model that
is doing it.

opw-4227241

closes odoo/enterprise#71751

X-original-commit: d193d73ac8dcafe6229135166519d8d5d48721b5
Related: odoo/odoo#183320
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>