[IMP] hr_recruitment: Report excluding duplicates
`Applicant Analysis` report was renamed to `Application Analysis`.
This PR also adds a filter which allows the user to filter out duplicate applications for a job position. A duplicate application is defined as an application to a job position where there is already another application with the same candidate_id.

Task-4222501

closes odoo/enterprise#71508

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>