[IMP] account_online_synchronization: Introduce webhooks between Odoo and OdooFin
Adds webhook event support on Odoofin.
With this addition, we can now update user's database in real-time when relevan, therefore
drastically reduce the cases where Odoo and OdooFin are not synchronized or when data are
not up-to-date.

For now the only supported type is `update_auth_exp_date` which notify the Odoo clients
that they need to fetch their authorization expiring date.

task-3822526

closes odoo/enterprise#71730

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>