[FIX] planning, social, web_gantt: fix `CP buttons` extra spacing
This commit removes an extra spacing added on top of the gap rule already
defined at the control panel level. As this extra spacing was leading to
a weird and inconsistent visual result, we simply remove it.

It also fixes an issue about `GanttView.Buttons` always being present
in the DOM even if empty, leading to an extra spacing due to the
`d-flex gap-1` set on the parent. To handle that scenario, we simply add
a `d-empty-none` rule on it.

task-3833933

closes odoo/enterprise#71622

Related: odoo/odoo#183077
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>