[FIX] whatsapp: send audio and text separately
The Whatsapp API doesn't allow to send captions with audio.
After this commit audio and body text will be sent separately.
If an additional attachment is set, it will be sent with the body.

We consider the `attachments` parameter is not used by discuss hence there is no
explicit support for automatically splitting the message there.

task-4007616

closes odoo/enterprise#77885

closes odoo/enterprise#78470

X-original-commit: 9d6210c42f39b3f3e9d8430ac78ffbf14c41483a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Renaud Thiry <reth@odoo.com>