[IMP] sale_subscription: display negative discount for upsell orders
**Before this PR:**
Previously, in the upsell order portal view, if the discount was less than 0, 
the discount field was hidden. This led to confusion as users could not see the 
applied discounts.

**After this PR:**
This improvement enables the discount field to be visible for upsell orders even
if the discount is negative. The change ensures that users have clearer information 
because all discounts, positive or negative, are displayed consistently.

task-4033128

closes odoo/enterprise#66833

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Co-authored-by: Arnaud Joset <arj@odoo.com>
Co-authored-by: Gurpreet Singh <gusi@odoo.com>