[FIX] stock_barcode: prefill transfers returns
Commit odoo/odoo@57b8b2487def4 make the return wizard fill the product
quantity to 0 default. This break the return flow in the stock_barcode
application as the return picking is created on the fly but empty so an
error is raised.

To encounter this, returns created via stock_barcode app have their
quantities set to the original stock move ones. The return picking is
thus fill according to the need.

closes odoo/enterprise#71693

X-original-commit: 46dffe709833e75972d3cf3f2b9071bab7c677b4
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: William Henrotin (whe) <whe@odoo.com>