[FIX] account_{batch_payment,sepa,sepa_direct_debit}: fix sanitize_communication
This commit aims to fix an issue in the sanitize_communication function
and move its logic to the account_batch_payment module to improve reusability
and testability.

Problem with the sanitize_communication function:
The function trims the communication text after removing the '/' characters.
However, if the communication text contains '/' and is truncated, the '/'
could end up at the end of the string (see tests for more examples).
The solution is to remove the '/' after truncating the text.

Additionally, the sanitize_communication function has been moved to the
account_batch_payment module, since this module is a dependency for both
account_sepa and account_sepa_direct_debit, these two modules can now
use it as well.

The functions sanitize_communication and _sanitize_communication have
been marked as deprecated and will be removed in the master branch.

opw-4220421

closes odoo/enterprise#73038

X-original-commit: fc735daadb8691f1e057e62bd65e9cf39bbe879a
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>