[IMP] l10n_be_hr_payroll: consecutive sick leaves without certificate
In Belgium, it is not legal to take two consecutive sick leaves
without certificate. For the leaves not to be considered
consecutive there should be at least one working day between them,
and so public holidays and non-working days would not be considered.

This commit adds an option to manage consecutive leaves and
adds a leave type for sick days without certificate so that this
rule could be enforced.

task-4213253

closes odoo/enterprise#71557

Related: odoo/odoo#187868
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>