[IMP] delivery: display shipping pickup point name
Issue: When a sale order made from the website with a pickup point
(for instance using a sendcloud delivery method) is confirmed, a new
partner is created to mix the data of the partner making the order and
the address of the pickup point. If the partner is named Bob, this new
partner will be named Bob, Bob. It woul dbe much clearer if he was named:
Bob, pickup point name.

opw-4181787

closes odoo/enterprise#71944

X-original-commit: faa162e5e40c9f6fb9ee4f57b9ec2975afff6fda
Related: odoo/odoo#183642
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>