[IMP] point_of_sale,*: revamp order pos_reference
In this commit, we are making the order numbers to be sequential across a
session. To accomplish this, we introduce `pos.session.order_seq_id` field which
points to an `ir.sequence` record. Each time an order is to be created, we first
request for a set of `pos.order` references namely `pos_reference`,
`sequence_number` an `tracking_number` that are derived from the mentioned field.

`pos_reference` is now in the form: YYLL-SSS-FOOOO where:
    * YY is the current year
    * LL is the login number (which serves as proxy to the device)
    * SSS is the session id
    * F is 1 or 0 (1 if sequence number is generated from client)
    * OOOO is the sequence number.

As a result, orders created in the same session will be assigned serialized
`sequence_number`. Deviation from the sequence (happens when order is started
from offline mode) will be explicit because we assigned `F=1` (see
`pos_reference`) and the `sequence_number` is negative.

We are also replacing the `login_number` field with a sequence
(`pos.session.login_number_seq_id`) to avoid error from concurrent update of the
`loign_number`.

closes odoo/enterprise#71632

Task-id: 4141140
Related: odoo/odoo#183085
Related: odoo/upgrade#6603
Signed-off-by: David Monnom (moda) <moda@odoo.com>