[IMP] marketing_automation: support dynamic domains
Allow marketing activities to use dynamic domains to be evaluated
at the time the domain is used. This allows the front-end filters
to be fully used in the automated flows to support flows such as
executing actions only for customers who have ordered something
"within 2 weeks" without having to update and relaunch a campaign
manually.

Domains are merged as expected using AST manipulations and stringification.
Details for this are available on the matching community commit
in mass mailing filters.

task-4240689

Part-of: odoo/enterprise#72570
Related: odoo/odoo#184830
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>