[FIX] documents: disable dragging in the TRASH folder
Before this commit:

Dragging and dropping documents from the TRASH to another folder will result in
a traceback or user error.
Dragging and dropping documents within the TRASH will trigger an unusual notification.

After this commit:

Disable dragging in the TRASH folder and inform the user with a valid reason.

Task-4231709

closes odoo/enterprise#73470

X-original-commit: 5ac0a23f9a4c8b9e287695b69fde279e194a386f
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>