[IMP] account_reports: Hide tax_period filter when not needed
Before, the tax period filter was always shown even if the dates could
be made from the other filters

Now we hide it if the periodicity is monthly or quarterly and that the
start date is the first day of the year. Otherwise we can show it
because it will create different dates not possible with the default
filters unless we use the custom dates filter.

task-4197475

closes odoo/enterprise#72222

X-original-commit: 2af26c034bb69e3df41ca37960bcdd2fa66c8671
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>