[FIX] documents: fix document_delete_tour failure
The tour was failing because the test folder ("Folder1") was not always
selected when starting the test with its access token.

The cause (very likely, see in depth explanation below) was that the test was
started with the URL of the "Folder1" through the call of start_tour in python
but another URL was specified in the tour definition in js ("/odoo").

We solve the problem by removing the start URL in the tour definition in js
(and also remove the part of the test that was launching the document app as
the start URL launch the document app directly).

In depth explanation:

Logging the start URL in the function startRouter has shown that the test first
call that function with the URL:
http://127.0.0.1:8069/web#action=...model=documents.document&...
then very soon after, call the startRouter again with the URL:
http://127.0.0.1:8069/odoo

In the best scenario, the first start has the time to write the folder to be
restored by the second call in the local storage . But the test failure, where
the wrong folder was selected, are likely due to a lack of time of the first
call to write the folder to restore in the local storage.

Task-4247011

closes odoo/enterprise#73699

X-original-commit: 1fc91827fa496f45bd5a37a86c8d2df88780cc2b
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>