[IMP] project_enterprise: allow resequencing in map view of Project/FSM
The JS framework introduced the option to reorder records from the map view side panel.
This is particularly useful to optimize the itinerary.
However, this feature is only available when records are ordered by an integer field, which isn't the case at the moment in project/fsm.
As a consequence, we need to order tasks by 'sequence > start date' in the map view of project/fsm.

task-4245177

Part-of: odoo/enterprise#71689
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>