[IMP] partner_commission: group POs coming from different salespeople
Previously, commission POs would be grouped per salesperson. In large
companies commission is generated through different salespeople which
leads to many POs to the same partner. This causes more overhead for
the accounting teams on both sides: there are more POs to process on
one side, and more invoices to process on the other.

To minimize the number of POs, we remove the salesperson from the
domain and stop assigning them as user_id (there can now be
multiple). To ensure they keep access to the same POs as before, we
add them as a follower on the PO and modify the record rule to take
this into account.

task-4250635

closes odoo/enterprise#71715

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>