[FIX] mrp_workorder: always select first option
Depending of the selected work centers in the Shop Floor view, different
work center buttons will be visible ("All Mo", "My WO", "Assembly 1",
...) When no work center are selected by default, the first one should
be the selected one, but for now, in this case it's always "All MO" who
is selected.
This commit fixes that, the first button/WC will always the selected one
if there is no WC id in the context.

Part-of: odoo/enterprise#74116
Related: odoo/odoo#187767
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>