[FIX] mrp_workorder: await the catalog update before reloading records
Issue:

Since 18.0, the catalog is used in certain actions of the shopfloor
However, if you leave the catalog fast enough after your last change
(something of the order of ~0.5 sec in localhost), you can trigger
a reload of the shopfloor records and call a `web_read` prior to
the update of these records.

Note:

This is easily reproducible by hand in localhost and should worsen if
you have server delay. In addition, it makes it impossible to write a
proper tour involving the catalog in the shopfloor.

Steps to reproduce:

- Create a product with a bom and an operation op1
- Create and confirm an MO for 1 unit of that product
- Go to the shopfloor > on the operation > wheel > Add component
- Add a component to be consumed in that operation
- Close the catalog fast (using ESC for instance)

> While the componenet was correctly added to the MO and linked to
> the WO it is not visible on the shopfloor.

Cause of the issue:

When a product is added from the catalog it triggers an rpc call to
update the data's of the MO (notably creating a new raw move):
https://github.com/odoo/odoo/blob/193c9a49dfc039ee93fbc8e171819236697b5c50/addons/product/static/src/product_catalog/kanban_record.js#L62-L69
https://github.com/odoo/odoo/blob/193c9a49dfc039ee93fbc8e171819236697b5c50/addons/mrp/models/mrp_production.py#L2898-L2901
On the other hand, closing the product catalog will reload the mrp
record that started the action that opened the catalog:
https://github.com/odoo/enterprise/blob/c1346dd2279bf3882b47b260b8d03ef9651ffaf1/mrp_workorder/static/src/mrp_display/dialog/mrp_menu_dialog.js#L41-L45
Since this reload is currently not waiting for the update of the record
the associated `web_read` might be called prior to the last update of
the record and the newly created raw moves might not be displayed.

Fix:

We use the props option of the action opening the catalog to transfer a
callback to reload the records when all the record update promisses are
resolved.

To be more precise:

The `ProductCatalogKanbanController` component contains a list of
promises (initially empty): `catalogKanbanUpdates` aswell as an optional
callback props `onCatalogUpdated`. This list is going to be populated
each time that a `ProductCatalogKanbanRecord` make a quantity update
(used stock moves on the mrp.production). To populate this array of
promises a method is defined in the controller:
`pushCatalogKanbanUpdate`. This method needs to be called by the
`KanbanRecords` an hence is transported using xpath from the
`ProductCatalogKanbanController` to the `ProductCatalogKanbanRecord`
via the `ProductCatalogKanbanRenderer`. Once the controller is detroyed
(and hence each kanban record that had to update a quantity started its
call) if a callback was given to the props of the Kanban controller we
wait for all the update promises to be resolved and we call that call
back.

opw-4199156

closes odoo/enterprise#79986

X-original-commit: eedd23a703ac4a4f98570b02a63f94209ecdbffd
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>