[FIX] project_enterprise: grey out weekends project gantt groupby
Steps to reproduce:
- Any project > Gantt view > Add any groupby filter

The weekends are no longer greyed out like they were without the filter.
This happens because _gantt_unavailablities in this module defers calls
to the super() method which is empty unless the field is user_id. This
makes sense since we don't want to get unavailability by employee if the
tasks are not sorted by user, we could however use company leaves
instead of leaving every day as available under other filters.

opw-4237953

closes odoo/enterprise#71999

X-original-commit: 4611d84e6cb51082edec28ae86fa3d71845016a1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>