[IMP] account_reports: Allow grouping by related non-stored fields
At the moment, the allowed groupbys for the standard report engines
(account codes, domain, tax tags) need to be stored fields on
`account.move.line`, or be defined in the custom handler.

We want to be able to group by any field that is searchable and has
`_field_to_sql` implemented for it. For example, for consolidation,
users might want to create a non-stored field related to
`account_id.code` and group by it.

After this commit, groupbys can be any searchable field for which
`_field_to_sql` is defined, regardless of whether it is stored or not.

task-4259552

closes odoo/enterprise#78520

X-original-commit: 11f0425be31622e23f4f4d47b9620398ac0b8338
Signed-off-by: William André (wan) <wan@odoo.com>