[IMP] l10n_br_edi: display calculated CFOP on invoice lines
Avatax determines the CFOP [1] of each invoice line during tax
calculation. It bases it on the chosen operation type, product,
contact and company configuration. The CFOP values are submitted
during electronic invoicing.

Displaying them to the user before electronic invoicing gives the user
a chance to detect misconfiguration earlier, and correct it without
having to deal with reversing an electronic invoice.

This commit updates the translations to include these new terms, and
also cleans them up so they match the code again [2].

[1] Código Fiscal de Operações e de Prestações: a fiscal code defined
by the Brazilian authorities to describe the type of business
transaction.
[2] Went out of sync due to 402326b64c14faf1658fc1c42ca19218df4c401d
not updating the translations.

task-4250764

closes odoo/enterprise#72247

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>