[IMP] hr_appraisal: introduce a wizard to create multiple appraisals
Appraisals could be created for multiple employees from the list view
in the Employees App. However, this feature isn't present in the
Appraisals app.

This feature should be directly accessible from the Appraisals app.
Hence, a new wizard is created to ease that process. The same wizard
will also be used to create multiple appraisals from the Employees app.

task-4191714

closes odoo/enterprise#71291

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>