[IMP] account_{followup,reports}: Customer Statements responsible partner
Partner Ledgers can be sent to Partners by email which is usually signed by the `followup_responsible`
user, however, `_get_followup_responsible` is defined in `account_followup` which may not be installed.

To solve this dependency issue, we define the method in `account_reports`, and the method in
`account_followup` now extends it.

No task - issue identified by marketing team upon review of the email template.

closes odoo/enterprise#71950

X-original-commit: 806b80b55284a764dc148953dd44ba0aec7fd41d
Signed-off-by: William André (wan) <wan@odoo.com>