[FIX] approvals: fetch correct employee for request
Problem: In a multi-company environment where a user has an employee record per company,
the approval request will retrieve the incorrect employee record upon confirmation,
which leads to an access error.

Purpose: Upon confirming the approval request,
the employee record from the same company as the
request should be fetched, rather than retrieving
the latest employee record made for the user.

Steps to reproduce on Runbot:

1. Install Approvals and Employees
2. Create a default company
3. Create a non-default company
4. Create a Test Employee user on the default company,
enable multi-company rights,
and create an employee record for the default company.
Switch to the non-default company and create an employee record for the user.
5. Repeat Step 4 when creating a Test Manager user
6. On the employee record for Test Employee,
set Test Manager as the manager on both companies.
7. Create an approval type on the default company
and set "Manager is required approver"
8. Login as Test Employee and create an approval request
for the approval type created in Step 7 on the default company
9. Submit the request and an AccessError occurs

Note: Bug was not reproducible on previous versions because the record rule
for Employee multi company rule for Employees module are different.

opw-4212341

closes odoo/enterprise#74724

X-original-commit: 0e6af91622afbeacd70c29ee9d20656b4ae64562
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>