[FIX] account_invoice_extract: fix traceback when uploading a vendor bill
Currently, a traceback may occur when the user tries to upload a vendor bill
Where the invoice has an Indian company and a rounding error.

Error:- "KeyError: 'groups_by_subtotal'"

This is because of the major refactor done from the below commit

https://github.com/odoo/odoo/pull/180062/commits/73d585a85b093db07e4eb5541aaa03e74af5fd62#diff-1ee3fce434db0c4e897973eebf5c1be501196cbd413e6c250ecc973238f939b9

Initially, we have a `groups_by_subtotal` in tax_totals, but after the refactor
we return subtotals containing tax_groups.

The code has been adjusted according to the above commit to resolve this issue.

sentry-5963987892

closes odoo/enterprise#74090

X-original-commit: 64ac515410de5891aeaa6b2a4eeafe2dcc858e95
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>