[IMP] sign: improve validation for unsigned documents
Before this commit, if a document wasn't fully signed, the system redirected to
a new page and triggered a UserError, but the user couldn't see the error
message.

With this commit, the validation is improved to properly show a clear error
message when documents are not fully signed, ensuring the user is informed.

task-4268726

closes odoo/enterprise#72282

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>