[IMP] industry_fsm_report: add test to check action_send_report in task
Before this commit, no test exists to check if the `action_send_report`
method in `project.task` model returns the right action.

This commit adds a test to check all the cases managed inside the
`action_send_report` method to make sure the behavior of that method is
fully tested. This commit also checks the methods overriden in
`industry_fsm` and `industry_fsm_report` for `base.document.layout`
wizard to make sure the template rendered is the one defined in
`industry_fsm` module.

task-4237400

closes odoo/enterprise#71962

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>